+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## next
 
 - Breaking: Added new configuration variables to mumble-authenticator.ini.
+- Breaking: Changed database schema: `ALTER TABLE user ADD account_active TINYINT DEFAULT 1 NULL;`
+- Fix: Accounts are now disabled if the player does not have any group required by the plugin.
 - Fix: Allow SuperUser login.
 
 ## 0.4.0, 2023-01-28